proyecto/requirements.txt
django-crispy-forms
cx_Oracle
pillow